requirements.txt
langchain
langchain-community
langchain-openai
langchain_ollama
langchain-chroma
ragas
langchain-text-splitters
langsmith
python-dotenv
pypdf
rapidfuzz
python-magic
beautifulsoup4
requests
ipykernel
ipywidgets
openai
spacy
langchain-qdrant
fastembed
unstructured
streamlit
langchain-qdrant